requirements.txt
python-dotenv>=0.19.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
jsonschema>=4.17.0
dataclasses-json>=0.6.0
graphviz>=0.20.0
chromadb==0.5.23
tokenizers==0.20.3
transformers==4.45.0
sentence-transformers>=2.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
Pillow>=10.0.0
requests>=2.31.0